requirements.txt
Django==5.0.3
django-ckeditor
fontawesomefree